테스트 코드도 패키지 세분화를 적용하라
