+++ app-server/app/src/test/java/com/codesoom/myseat/controllers/login/LoginControllerTest.java
@@ -1,10 +1,10 @@
-package com.codesoom.myseat.controllers;
+package com.codesoom.myseat.controllers.login;
 
-import com.codesoom.myseat.controllers.seats.SeatAddController;
-import com.codesoom.myseat.domain.Seat;
-import com.codesoom.myseat.dto.SeatAddRequest;
-import com.codesoom.myseat.services.AuthenticationService;
-import com.codesoom.myseat.services.SeatAddService;
+import com.codesoom.myseat.controllers.login.LoginController;
+import com.codesoom.myseat.domain.User;
+import com.codesoom.myseat.dto.LoginRequest;
+import com.codesoom.myseat.services.auth.AuthenticationService;
+import com.codesoom.myseat.services.users.UserService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
@@ -15,37 +15,35 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
 import static org.mockito.BDDMockito.given;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(SeatAddController.class)
+@WebMvcTest(LoginController.class)
 @AutoConfigureRestDocs(
         uriScheme = "https", 
         uriHost = "api.codesoom-myseat.site"
 )
-class SeatAddControllerTest {
-    private static final String ACCESS_TOKEN
+class LoginControllerTest {
+    private static final Long ID = 1L;
+    private static final String NAME = "테스터";
+    private static final String EMAIL = "test@example.com";
+    private static final String PASSWORD = "1111";
+    private static final String ENCODED_PASSWORD
+            = "$2a$10$hxqWrlGa7SQcCEGURjmuQup4J9kN6qnfr4n7j7R3LvzHEoEOUTWeW";
+
+    private static final String ACCESS_TOKEN 
             = "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjF9.ZZ3CUl0jxeLGvQ1Js5nG2Ty5qGTlqai5ubDMXZOdaDk";
 
-    private static final String INVALID_TOKEN
-            = "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjF9.ZZ3CUl0jxeLGvQ1Js5nG2Ty5qGTlqai5ubDMXZOdaDk2";
-    
-    private static final Long ID = 1L;
-    private static final int NUMBER = 3;
-
-    private static final String EMAIL = "test@example.com";
-    
     @Autowired
     private MockMvc mockMvc;
 
@@ -53,41 +51,51 @@ class SeatAddControllerTest {
     private AuthenticationService authService;
 
     @MockBean
-    private SeatAddService seatAddService;
+    private UserService userService;
 
     @Test
-    @WithMockUser
-    @DisplayName("POST /seat responses")
-    void seat_add_success() throws Exception {
-        // given
-        SeatAddRequest request = SeatAddRequest.builder()
-                .number(NUMBER)
+    @DisplayName("로그인 성공")
+    void login_success() throws Exception {
+        //given
+        LoginRequest request = LoginRequest.builder()
+                .email(EMAIL)
+                .password(PASSWORD)
                 .build();
 
-        given(authService.parseToken(ACCESS_TOKEN))
-                .willReturn(1L);
-        
-        given(seatAddService.createSeat(NUMBER))
-                .will(invocation -> Seat.builder()
-                        .id(ID)
-                        .number(NUMBER)
-                        .build());
+        User user = User.builder()
+                .id(ID)
+                .name(NAME)
+                .email(EMAIL)
+                .password(ENCODED_PASSWORD)
+                .build();
+
+        given(userService.findByEmail(EMAIL))
+                .willReturn(user);
+
+        given(authService.login(user, PASSWORD))
+                .willReturn(ACCESS_TOKEN);
 
         // when
-        ResultActions subject = mockMvc.perform(post("/seat")
-                .header("Authorization", "Bearer " + ACCESS_TOKEN)
+        ResultActions subject = mockMvc.perform(post("/login")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(toJson(request)));
 
         // then
-        subject.andExpect(status().isNoContent());
+        subject.andExpect(status().isOk())
+                .andExpect(jsonPath("$.accessToken").value(ACCESS_TOKEN))
+                .andExpect(jsonPath("$.name").value(NAME));
 
         // docs
-        subject.andDo(document("seat-add",
+        subject.andDo(document("login",
                 preprocessRequest(prettyPrint()),
                 preprocessResponse(prettyPrint()),
                 requestFields(
-                        fieldWithPath("number").type(JsonFieldType.NUMBER).description("좌석 번호")
+                        fieldWithPath("email").type(JsonFieldType.STRING).description("이메일"),
+                        fieldWithPath("password").type(JsonFieldType.STRING).description("비밀번호")
+                ),
+                responseFields(
+                        fieldWithPath("accessToken").type(JsonFieldType.STRING).description("토큰"),
+                        fieldWithPath("name").type(JsonFieldType.STRING).description("회원 이름")
                 )
         ));
     }
