+++ app-server/app/src/test/java/com/codesoom/myseat/services/auth/AuthenticationServiceTest.java
@@ -1,8 +1,9 @@
-package com.codesoom.myseat.services;
+package com.codesoom.myseat.services.auth;
 
 import com.codesoom.myseat.domain.User;
 import com.codesoom.myseat.repositories.RoleRepository;
 import com.codesoom.myseat.repositories.UserRepository;
+import com.codesoom.myseat.services.auth.AuthenticationService;
 import com.codesoom.myseat.utils.JwtUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
