+++ app-server/app/src/test/java/com/codesoom/myseat/services/reservations/retrospectives/RetrospectiveServiceTest.java
@@ -1,4 +1,4 @@
-package com.codesoom.myseat.services;
+package com.codesoom.myseat.services.reservations.retrospectives;
 
 import com.codesoom.myseat.domain.Reservation;
 import com.codesoom.myseat.domain.Retrospective;
@@ -7,6 +7,7 @@ import com.codesoom.myseat.exceptions.NotRegisteredReservation;
 import com.codesoom.myseat.repositories.ReservationRepository;
 import com.codesoom.myseat.repositories.RetrospectiveRepository;
 import com.codesoom.myseat.repositories.UserRepository;
+import com.codesoom.myseat.services.reservations.retrospectives.RetrospectiveService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
