+++ app-server/app/src/test/java/com/codesoom/myseat/services/seats/SeatListServiceTest.java
@@ -1,7 +1,8 @@
-package com.codesoom.myseat.services;
+package com.codesoom.myseat.services.seats;
 
 import com.codesoom.myseat.domain.Seat;
 import com.codesoom.myseat.repositories.SeatRepository;
+import com.codesoom.myseat.services.seats.SeatListService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
